finish first round of refactor and clean up
